requirements.txt
torch==1.10.2
torchvision==0.11.3
Pillow==7.0.0
numpy==1.21.0
opencv-python==4.2.0.34
pyrebase
pyqt5